feat(dedup): go live
Refs: #2569
