+++ packages/lambdas/src/fhir-to-medical-record.ts
@@ -6,7 +6,6 @@ import {
 import { getFeatureFlagValueStringArray } from "@metriport/core/external/aws/app-config";
 import { Brief, bundleToBrief } from "@metriport/core/external/aws/lambda-logic/bundle-to-brief";
 import { bundleToHtml } from "@metriport/core/external/aws/lambda-logic/bundle-to-html";
-import { bundleToHtml as bundleToHtmlNoDedup } from "@metriport/core/external/aws/lambda-logic/bundle-to-html-no-dedup";
 import { bundleToHtmlADHD } from "@metriport/core/external/aws/lambda-logic/bundle-to-html-adhd";
 import {
   getSignedUrl as coreGetSignedUrl,
@@ -16,17 +15,8 @@ import {
 import { getEnvType } from "@metriport/core/util/env-var";
 import { out } from "@metriport/core/util/log";
 import { uuidv7 } from "@metriport/core/util/uuid-v7";
-import {
-  isFhirDeduplicationEnabledForCx,
-  isAiBriefFeatureFlagEnabledForCx,
-} from "@metriport/core/external/aws/app-config";
-import { EventTypes, analytics } from "@metriport/core/external/analytics/posthog";
+import { isAiBriefFeatureFlagEnabledForCx } from "@metriport/core/external/aws/app-config";
 import { errorToString, MetriportError } from "@metriport/shared";
-import { elapsedTimeFromNow } from "@metriport/shared/common/date";
-import { deduplicateFhir } from "@metriport/core/fhir-deduplication/deduplicate-fhir";
-import { SearchSetBundle } from "@metriport/shared/medical";
-import { Resource } from "@medplum/fhirtypes";
-import { uploadConsolidatedBundleToS3 } from "@metriport/core/command/consolidated/consolidated-on-s3";
 import chromium from "@sparticuz/chromium";
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
@@ -80,58 +70,6 @@ export async function handler({
     const isADHDFeatureFlagEnabled = cxsWithADHDFeatureFlagValue.includes(cxId);
     const bundle = await getBundleFromS3(fhirFileName);
     const isBriefFeatureFlagEnabled = await isAiBriefEnabled(generateAiBrief, cxId);
-
-    const dedupEnabled = await isFhirDeduplicationEnabledForCx(cxId);
-    if (dedupEnabled) {
-      const deduplicatedBundle = deduplicateSearchSetBundle(bundle);
-      await uploadConsolidatedBundleToS3({
-        patient: {
-          id: patientId,
-          cxId: cxId,
-        },
-        bundle: deduplicatedBundle,
-        s3BucketName: bucketName,
-        dedupEnabled,
-      });
-
-      const initialBundleLength = bundle.entry?.length;
-      const startedAt = new Date();
-
-      const finalBundleLength = deduplicatedBundle.entry?.length;
-
-      const deduplicationAnalyticsProps = {
-        distinctId: cxId,
-        event: EventTypes.fhirDeduplication,
-        properties: {
-          patientId: patientId,
-          initialBundleLength,
-          finalBundleLength,
-          duration: elapsedTimeFromNow(startedAt),
-        },
-      };
-      analytics(deduplicationAnalyticsProps);
-
-      const aiBriefContent = isBriefFeatureFlagEnabled
-        ? await bundleToBrief(deduplicatedBundle, cxId, patientId)
-        : undefined;
-
-      const briefFileName = createMRSummaryBriefFileName(cxId, patientId, dedupEnabled);
-      const aiBrief = prepareBriefToBundle({ aiBrief: aiBriefContent });
-
-      const html = bundleToHtmlNoDedup(deduplicatedBundle, aiBrief);
-      const hasContents = doesMrSummaryHaveContents(html);
-      log(`MR Summary has contents: ${hasContents}`);
-      const htmlFileName = createMRSummaryFileName(cxId, patientId, "html", dedupEnabled);
-
-      await storeMrSummaryAndBriefInS3({
-        bucketName,
-        htmlFileName,
-        briefFileName,
-        html,
-        aiBrief: aiBriefContent,
-        log,
-      });
-    }
 
     // TODO: Condense this functionality under a single function and put it on `@metriport/core`, so this can be used both here, and on the Lambda.
     const aiBriefContent = isBriefFeatureFlagEnabled
@@ -436,14 +374,4 @@ async function createFeedbackForBrief({
       },
     });
   }
-}
-
-function deduplicateSearchSetBundle(
-  fhirBundle: SearchSetBundle<Resource>
-): SearchSetBundle<Resource> {
-  const deduplicatedBundle = deduplicateFhir(fhirBundle);
-  return {
-    ...deduplicatedBundle,
-    type: "searchset",
-  };
 }
